feat(settings): add `Desktop & Dock` tab to settings, and move top bar and dock settings to this new tab
