+++ src/_locales/cmn-TW.yml
@@ -32,6 +32,7 @@ settings:
 
   # Settings menu
   menu_general: 一般
+  menu_desktop_and_dock: 桌面與 Dock
   menu_appearance: 外觀
   menu_home: 首頁
   menu_search_page: 搜尋頁
