+++ src/_locales/jyut.yml
@@ -32,6 +32,7 @@ settings:
 
   # Settings menu
   menu_general: 一般
+  menu_desktop_and_dock: 桌面同 Dock
   menu_appearance: 版面
   menu_home: 主頁
   menu_search_page: 搵嘢頁
