+++ src/components/Settings/Settings.vue
@@ -12,6 +12,7 @@ const { t } = useI18n()
 
 const settingsMenu = {
   [MenuType.General]: defineAsyncComponent(() => import('./components/General.vue')),
+  [MenuType.DesktopAndDock]: defineAsyncComponent(() => import('./components/DesktopAndDock.vue')),
   [MenuType.Appearance]: defineAsyncComponent(() => import('./components/Appearance.vue')),
   [MenuType.SearchPage]: defineAsyncComponent(() => import('./components/SearchPage.vue')),
   [MenuType.Home]: defineAsyncComponent(() => import('./components/Home.vue')),
@@ -38,6 +39,12 @@ const settingsMenuItems = computed((): MenuItem[] => {
       icon: 'i-mingcute:settings-3-line',
       iconActivated: 'i-mingcute:settings-3-fill',
       title: t('settings.menu_general'),
+    },
+    {
+      value: MenuType.DesktopAndDock,
+      icon: 'i-mingcute:imac-line',
+      iconActivated: 'i-mingcute:imac-fill',
+      title: t('settings.menu_desktop_and_dock'),
     },
     {
       value: MenuType.Appearance,
@@ -134,7 +141,7 @@ function setCurrentTitle() {
         >
           <li v-for="menuItem in settingsMenuItems" :key="menuItem.value">
             <a
-              cursor-pointer w="40px group-hover:150px" h-40px
+              cursor-pointer w="40px group-hover:180px" h-40px
               rounded-30px flex items-center overflow-x-hidden
               duration-300 bg="hover:$bew-fill-2"
               :class="{ 'menu-item-activated': menuItem.value === activatedMenuItem }"
